added placeholder & fixed like counter
- refactored large like counter method and fixed it
- added a placeholder functionality for the username,
password and bio. This is done in the BasePanel. I
creted a new constructor which handles this logic.
As nearly every class uses the BasePanel constructor
I added the super keyword to all of them and adjusted
the newly created boolean values. If you want to change
this just cahnge the boolean values in the super call
of the respective class